refacto: split library file
